pyproject.toml
[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "health-funding-scraper"
version = "0.1.0"
description = "Agrégateur automatisé d'appels d'offres et de financements dans le domaine de la santé en France"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32",
    "beautifulsoup4>=4.12",
    "lxml>=5.3",
    "playwright>=1.49",
    "pandas>=2.2",
    "pydantic>=2.10",
    "pdfplumber>=0.11",
    "schedule>=1.2",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3", "pytest-asyncio>=0.24"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
